+++ src/ResourceManager/ServerManagement/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fix issue with Default Resource Group in CloudShell
 
 ## Version 4.1.0
 * Added Location Completer to -Location parameters allowing tab completion through valid Locations